+++ packages/stargate/src/fee.spec.ts
@@ -12,12 +12,49 @@ describe("GasPrice", () => {
     });
   });
 
-  it("can be constructed from a config string", () => {
-    const inputs = ["3.14", "3", "0.14"];
-    inputs.forEach((input) => {
-      const gasPrice = GasPrice.fromString(`${input}utest`);
-      expect(gasPrice.amount.toString()).toEqual(input);
-      expect(gasPrice.denom).toEqual("utest");
+  describe("fromString", () => {
+    it("works", () => {
+      const inputs: Record<string, { amount: string; denom: string }> = {
+        // Test amounts
+        "3.14utest": { amount: "3.14", denom: "utest" },
+        "3utest": { amount: "3", denom: "utest" },
+        "0.14utest": { amount: "0.14", denom: "utest" },
+        // Test denoms
+        "0.14sht": { amount: "0.14", denom: "sht" },
+        "0.14testtesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttest": {
+          amount: "0.14",
+          denom:
+            "testtesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttest",
+        },
+        "0.14ucoin2": { amount: "0.14", denom: "ucoin2" },
+        // eslint-disable-next-line @typescript-eslint/naming-convention
+        "0.14FOOBAR": { amount: "0.14", denom: "FOOBAR" },
+      };
+      for (const [input, expected] of Object.entries(inputs)) {
+        const gasPrice = GasPrice.fromString(input);
+        expect(gasPrice.amount.toString()).withContext(`Input: ${input}`).toEqual(expected.amount);
+        expect(gasPrice.denom).withContext(`Input: ${input}`).toEqual(expected.denom);
+      }
+    });
+
+    it("errors for invalid gas price", () => {
+      // Checks basic format <amount><denom>
+      expect(() => GasPrice.fromString("")).toThrowError(/Invalid gas price string/i);
+      expect(() => GasPrice.fromString("utkn")).toThrowError(/Invalid gas price string/i);
+      expect(() => GasPrice.fromString("@utkn")).toThrowError(/Invalid gas price string/i);
+      expect(() => GasPrice.fromString("234")).toThrowError(/Invalid gas price string/i);
+      expect(() => GasPrice.fromString("-234tkn")).toThrowError(/Invalid gas price string/i);
+      // Checks details of <denom>
+      expect(() => GasPrice.fromString("234t")).toThrowError(/denom must be between 3 and 128 characters/i);
+      expect(() => GasPrice.fromString("234tt")).toThrowError(/denom must be between 3 and 128 characters/i);
+      expect(() =>
+        GasPrice.fromString(
+          "234ttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttt",
+        ),
+      ).toThrowError(/denom must be between 3 and 128 characters/i);
+      // Checks details of <amount>
+      expect(() => GasPrice.fromString("3.utkn")).toThrowError(/Fractional part missing/i);
+      expect(() => GasPrice.fromString("..utkn")).toThrowError(/More than one separator found/i);
     });
   });
 });