+++ CHANGELOG.md
@@ -12,8 +12,17 @@ and this project adheres to
   `CosmWasmClient.broadcastTx` and `StargateClient.broadcastTx` ([#800]). This
   bug was introduced with the switch from broadcast mode "commit" to "sync" in
   version 0.25.0.
+- @cosmjs/launchpad, @cosmjs/stargate: Avoid the use of named capture groups in
+  `GasPrice.fromString` to restore ES2017 compatibility and make the library
+  work with Hermes ([#801]; thanks [@AlexBHarley]).
+- @cosmjs/launchpad: Adapt `GasPrice.fromString` denom pattern to Cosmos SDK
+  0.39 rules: reduce denom length to 16 and allow digits in denom.
+- @cosmjs/stargate: Adapt `GasPrice.fromString` denom pattern to Cosmos SDK 0.42
+  rules: allow lengths up to 128, allow upper case letters and digits.
 
 [#800]: https://github.com/cosmos/cosmjs/issues/800
+[#801]: https://github.com/cosmos/cosmjs/issues/801
+[@alexbharley]: https://github.com/AlexBHarley
 
 ## [0.25.2] - 2021-05-11